Style Changes
Main.h removed.

Major program control flow loops moved from Main.cpp to new file
EventLoops.cpp. EventLoops.h header file also created to match.
