+++ EventLoops.h
@@ -1,7 +1,7 @@
 //Samuel Weber
 
-#ifndef MAIN_H
-  #define MAIN_H
+#ifndef EVENTLOOPS_H
+  #define EVENTLOOPS_H
 
   #include "Draw.h"
   #include "Keybind.h"
@@ -9,7 +9,7 @@
   #include "Party.h"
   #include "Terr.h"
 
-  //Main.cpp
+  //EventLoops.cpp
   bool loopAnyState(SDL_Renderer *ren, SDL_Event &e, Terr *terr, Party *party, Sprite *npc, gameState &state);
   bool loopBattle(SDL_Event &e);
   bool loopMap(SDL_Renderer *ren, SDL_Texture *tiles, Terr *terr, SDL_Event &e, Party *party);
